# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Focusly Backend API",
    "description": "Local AI Learning Platform converting documents into summaries, quizzes, and narrated videos.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/admin/admin/status": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get System Status",
        "description": "Returns system status and connection state for admin.",
        "operationId": "get_system_status_api_admin_admin_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/status": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get System Status",
        "description": "Returns system status and connection state for admin.",
        "operationId": "get_system_status_api_admin_status_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List All Users Admin",
        "description": "Lists all platform users along with their monthly usage statistics.\nRequires user.is_admin flag.",
        "operationId": "list_all_users_admin_api_admin_users_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/users/{user_id}": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get User Details Admin",
        "description": "Retrieves detailed profile, status, and usage statistics for a specific user.",
        "operationId": "get_user_details_admin_api_admin_users__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/users/{user_id}/status": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Update User Status Admin",
        "description": "Suspends or activates a user account.",
        "operationId": "update_user_status_admin_api_admin_users__user_id__status_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/admin/usage": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get User Usage Stats",
        "description": "Returns aggregate usage statistics across all platform users.",
        "operationId": "get_user_usage_stats_api_admin_admin_usage_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/usage": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get User Usage Stats",
        "description": "Returns aggregate usage statistics across all platform users.",
        "operationId": "get_user_usage_stats_api_admin_usage_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/documents": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List Documents Admin",
        "description": "Lists documents across all users with optional filtering by user_id, status, and source_type.",
        "operationId": "list_documents_admin_api_admin_documents_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          },
          {
            "name": "source_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/documents/{document_id}": {
      "delete": {
        "tags": [
          "admin"
        ],
        "summary": "Admin Delete Document Override",
        "description": "Admin override to delete any document by ID.",
        "operationId": "admin_delete_document_override_api_admin_documents__document_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/logs": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "Get System Logs Admin",
        "description": "Returns recent system logs, connection events, and error records.",
        "operationId": "get_system_logs_admin_api_admin_logs_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "INFO",
              "title": "Level"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/maintenance/cleanup": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Run Maintenance Cleanup",
        "description": "Cleans expired share links and temporary upload files in /tmp/focusly_uploads.",
        "operationId": "run_maintenance_cleanup_api_admin_maintenance_cleanup_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/users/{user_id}/usage": {
      "put": {
        "tags": [
          "admin"
        ],
        "summary": "Adjust User Usage Limits",
        "description": "Allows admin to manually adjust user usage counters and monthly limit caps.",
        "operationId": "adjust_user_usage_limits_api_admin_users__user_id__usage_put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUsageAdjustRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/exams/research": {
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Research Exam Profile",
        "description": "Uses Gemini AI to research and generate structured exam profile:\nexam_name, sections, question_types, time_limit, total_questions, scoring_rules, disclaimer_text.\nRequires user.is_admin metadata flag.",
        "operationId": "research_exam_profile_api_admin_exams_research_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExamResearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExamProfileFull"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/exam-profiles": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List All Exam Profiles",
        "description": "Lists all researched exam profiles. Requires admin privileges.",
        "operationId": "list_all_exam_profiles_api_admin_exam_profiles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/admin/exams": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List All Exam Profiles",
        "description": "Lists all researched exam profiles. Requires admin privileges.",
        "operationId": "list_all_exam_profiles_api_admin_exams_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExamProfileFull"
                  },
                  "type": "array",
                  "title": "Response List All Exam Profiles Api Admin Exams Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/documents/{document_id}/status": {
      "get": {
        "summary": "Get Document Status Endpoint",
        "description": "Returns detailed document status including OCR confidence, virtual page map, progress, and error message.",
        "operationId": "get_document_status_endpoint_api_documents__document_id__status_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminUserStatusUpdate": {
        "properties": {
          "status": {
            "type": "string",
            "pattern": "^(active|suspended)$",
            "title": "Status"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "AdminUserStatusUpdate"
      },
...
      "ExamResearchRequest": {
        "properties": {
          "exam_name": {
            "type": "string",
            "title": "Exam Name"
          }
        },
        "type": "object",
        "required": [
          "exam_name"
        ],
        "title": "ExamResearchRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}